move contrib files to folder
