+++ nautobot_ssot/contrib/__init__.py
@@ -1,0 +1,3 @@
+"""SSoT Contrib."""
+from nautobot_ssot.contrib.adapter import NautobotAdapter
+from nautobot_ssot.contrib.model import NautobotModel
